refactor(Stdio): Refactor discovery
Allow better code reuse with `register`
